Minor fixes for types and adding tags as global extension

+++ src/OpenRpg.Combat/Extensions/CombatAbilityTemplateVariablesExtensions.cs
@@ -8,7 +8,7 @@ namespace OpenRpg.Combat.Extensions
     public static class CombatAbilityTemplateVariablesExtensions
     {
         public static float Cooldown(this AbilityTemplateVariables vars) => vars.GetFloatOrDefault(CombatAbilityTemplateVariableTypes.Cooldown, 0);
-        public static void Cooldown(this AbilityTemplateVariables vars, int value) => vars[CombatAbilityTemplateVariableTypes.Cooldown] = value;
+        public static void Cooldown(this AbilityTemplateVariables vars, float value) => vars[CombatAbilityTemplateVariableTypes.Cooldown] = value;
         
         public static Damage Damage(this AbilityTemplateVariables vars) => vars.GetAsOrDefault(CombatAbilityTemplateVariableTypes.Damage, () => new Damage(0, 0));
         public static void Damage(this AbilityTemplateVariables vars, Damage damage) => vars[CombatAbilityTemplateVariableTypes.Damage] = damage;
